Installment/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
chardet==5.2.0
diff-match-patch==20230430
Django==5.1.1
django-extensions==3.2.3
django-import-export==4.1.1
django-unfold==0.40.0
django-widget-tweaks==1.5.0
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pywin32-ctypes==0.2.3
reportlab==4.2.5
setuptools==75.1.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.2
